+++ src/main/java/wooteco/subway/aop/exception/ExceptionController.java
@@ -1,0 +1,18 @@
+package wooteco.subway.aop.exception;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import java.util.NoSuchElementException;
+
+@RestControllerAdvice({
+        "wooteco.subway.line.controller",
+        "wooteco.subway.station.controller"
+})
+public class ExceptionController {
+    @ExceptionHandler({IllegalArgumentException.class, NoSuchElementException.class})
+    public ResponseEntity<Void> exceptionHandler() {
+        return ResponseEntity.badRequest().build();
+    }
+}
